backend/requirements.txt
fastapi
uvicorn[standard]

typer[all]
requests

pymongo
redis
neo4j

python-dotenv
email-validator
typer[all]
requests
